Move field mappers one level up
