+++ lib/FieldMapper/ContentFieldMapper.php
@@ -6,7 +6,7 @@
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
-namespace EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper;
+namespace EzSystems\EzPlatformSolrSearchEngine\FieldMapper;
 
 use eZ\Publish\SPI\Persistence\Content as SPIContent;
 
@@ -15,7 +15,7 @@ use eZ\Publish\SPI\Persistence\Content as SPIContent;
  *
  * Content document field mapper maps Content to the search fields for Content document.
  */
-abstract class Content
+abstract class ContentFieldMapper
 {
     /**
      * Indicates if the mapper accepts the given $content for mapping.
