+++ lib/FieldMapper/ContentFieldMapper/Aggregate.php
@@ -6,25 +6,25 @@
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
-namespace EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\Location;
+namespace EzSystems\EzPlatformSolrSearchEngine\FieldMapper\ContentFieldMapper;
 
-use EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\Location as LocationMapper;
-use eZ\Publish\SPI\Persistence\Content\Location;
+use EzSystems\EzPlatformSolrSearchEngine\FieldMapper\ContentFieldMapper;
+use eZ\Publish\SPI\Persistence\Content;
 
 /**
- * Aggregate implementation of Location document field mapper.
+ * Aggregate implementation of Content document field mapper.
  */
-class Aggregate extends LocationMapper
+class Aggregate extends ContentFieldMapper
 {
     /**
      * An array of aggregated field mappers, sorted by priority.
      *
-     * @var \EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\Location[]
+     * @var \EzSystems\EzPlatformSolrSearchEngine\FieldMapper\ContentFieldMapper[]
      */
     protected $mappers = [];
 
     /**
-     * @param \EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\Location[] $mappers
+     * @param \EzSystems\EzPlatformSolrSearchEngine\FieldMapper\ContentFieldMapper[] $mappers
      *        An array of mappers, sorted by priority.
      */
     public function __construct(array $mappers = [])
@@ -37,25 +37,25 @@ class Aggregate extends LocationMapper
     /**
      * Adds given $mapper to the internal array as the next one in priority.
      *
-     * @param \EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\Location $mapper
+     * @param \EzSystems\EzPlatformSolrSearchEngine\FieldMapper\ContentFieldMapper $mapper
      */
-    public function addMapper(LocationMapper $mapper)
+    public function addMapper(ContentFieldMapper $mapper)
     {
         $this->mappers[] = $mapper;
     }
 
-    public function accept(Location $location)
+    public function accept(Content $content)
     {
         return true;
     }
 
-    public function mapFields(Location $location)
+    public function mapFields(Content $content)
     {
         $fields = [];
 
         foreach ($this->mappers as $mapper) {
-            if ($mapper->accept($location)) {
-                $fields = array_merge($fields, $mapper->mapFields($location));
+            if ($mapper->accept($content)) {
+                $fields = array_merge($fields, $mapper->mapFields($content));
             }
         }
