+++ lib/FieldMapper/LocationFieldMapper.php
@@ -6,7 +6,7 @@
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
-namespace EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper;
+namespace EzSystems\EzPlatformSolrSearchEngine\FieldMapper;
 
 use eZ\Publish\SPI\Persistence\Content\Location as SPILocation;
 
@@ -15,7 +15,7 @@ use eZ\Publish\SPI\Persistence\Content\Location as SPILocation;
  *
  * Location document field mapper maps Location to the search fields for Location document.
  */
-abstract class Location
+abstract class LocationFieldMapper
 {
     /**
      * Indicates if the mapper accepts given $location for mapping.
